requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
matplotlib>=3.7.0
pillow>=10.0.0
pycocotools>=2.0.6
numpy>=1.24.0
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.10.0
requests>=2.31.0
ultralytics>=8.0.0
# Waymo Open Dataset (수동 설치 필요)
# tensorflow>=2.12.0
# waymo-open-dataset-tf-2-11-0
